build game in project directory and load from there
This fixes a lot of problems I was having with trying to put custom game
components in the project directory and have them work with the
components that were built into the editor. This failed to work because
two instances of `bitecs` were loaded, each having different globals. So
the custom component would add itself to its own bitecs instance, and
the rest of the game couldn't see it.

This also, I think, solves the future problem of bundling the game with
an Electron app: With the entire game able to be bundled as one thing,
adding that to an Electron app becomes a matter of configuration.

This is extremely messy as it is, but it (mostly) works. However, we are
not yet generating the `bitwise.js` file. And we will need to regenerate
that file if any module changes.