fix editor not appearing if game reloaded while playing
If the editor is in "play" mode and the user modified some code, the
editor will reload in "play" mode. If the user then stops the "play"
game, the editor should appear again.

If the editor is initialized while it's hidden, the canvas has
no size. So the user clicks "Stop" to stop the playing game and then
sees a black canvas and has to reload to see the editor again.

Now we wait until the game is no longer playing before reinitializing
the editor.