fix problem adding system to scene
I feel like the ScenePanel has to know too much about the scene's data
model. Either we need to formalize that model (as a JSON schema, likely)
or add methods to the Scene class to work with the data. Since I already
tried just editing the Scene object and using freeze/thaw, it will
probably be the former.